More `provider_new` to `provider`; replacing the old provider
