--- provider/cmd/pulumi-resource-terraform/serve.go
@@ -1,36 +1,0 @@
-// Copyright 2016-2018, Pulumi Corporation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import (
-	"context"
-
-	"github.com/pulumi/pulumi/pkg/v3/resource/provider"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/util/cmdutil"
-	rpc "github.com/pulumi/pulumi/sdk/v3/proto/go"
-)
-
-// Serve launches the gRPC server for the Pulumi Terraform resource provider.
-func Serve(providerName, version string) {
-	// Start gRPC service.
-	err := provider.Main(
-		providerName, func(host *provider.HostClient) (rpc.ResourceProviderServer, error) {
-			return NewProvider(context.Background(), host, version), nil
-		})
-
-	if err != nil {
-		cmdutil.ExitError(err.Error())
-	}
-}
