+++ provider/pulumi-resource-command/schema.json
