+++ provider/shim/shim.go
@@ -2,75 +2,81 @@ package shim
 
 import (
 	"context"
-	"io/ioutil"
-	"log"
+	"encoding/json"
+	"fmt"
 
-	structpb "github.com/golang/protobuf/ptypes/struct"
 	"github.com/hashicorp/terraform-svchost/disco"
 	be "github.com/hashicorp/terraform/internal/backend"
 	backendInit "github.com/hashicorp/terraform/internal/backend/init"
-	pulumirpc "github.com/pulumi/pulumi/sdk/v3/proto/go"
+	"github.com/zclconf/go-cty/cty"
+	ctyjson "github.com/zclconf/go-cty/cty/json"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
-func InitTfBackend() {
-	backendInit.Init(disco.New())
+func InitTfBackend() { backendInit.Init(disco.New()) }
+
+type RemoteStateReferenceInputs struct {
+	// TODO: what is this for? Is it always be pulumi.String("remote")?
+	BackendType string
+
+	BackendConfig BackendConfig
+
+	// Workspace is a struct specifying which remote workspace(s) to use.
+	Workspaces WorkspaceStateArgs
 }
 
-func RemoteStateReferenceRead(ctx context.Context, req *pulumirpc.ReadRequest) (*pulumirpc.ReadResponse, error) {
-	// Prevent Terraform from logging minutia
-	log.SetOutput(ioutil.Discard)
+type BackendConfig struct {
+	// The name of the resource to read.
+	ResourceName string
 
-	// Pull the backendType out of the backendConfig, ensure it's a string
-	backendTypePB, hasBackendType := req.GetProperties().GetFields()["backendType"]
-	if !hasBackendType {
-		return nil, status.Errorf(codes.InvalidArgument,
-			"missing required property %q", "backendType")
-	}
-	if _, isString := backendTypePB.Kind.(*structpb.Value_StringValue); !isString {
-		return nil, status.Errorf(codes.InvalidArgument,
-			"expected a string value for property %q", "backendType")
-	}
-	backendType := backendTypePB.GetStringValue()
+	// Organization is the name of the organization containing the targeted workspace(s).
+	Organization string
+
+	// Hostname is the remote backend hostname to which to connect. Defaults to `app.terraform.io`.
+	Hostname string
+
+	// Token is the token used to authenticate with the remote backend.
+	Token string
+}
+
+// WorkspaceStateArgs specifies the configuration options for a workspace for use with the remote enhanced backend.
+type WorkspaceStateArgs struct {
+	// Name is the full name of one remote workspace. When configured, only the default workspace
+	// can be used. This option conflicts with prefix.
+	Name string
+
+	// Prefix is the prefix used in the names of one or more remote workspaces, all of which can be used
+	// with this configuration. If unset, only the default workspace can be used. This option
+	// conflicts with name
+	Prefix string
+}
+
+func RemoteStateReferenceRead(ctx context.Context, args RemoteStateReferenceInputs) (map[string]any, error) {
+	backendType := args.BackendType
 
 	// Ensure the backendType is known about by Terraform
 	backendInitFn := backendInit.Backend(backendType)
 	if backendInitFn == nil {
-		return nil, status.Errorf(codes.InvalidArgument,
-			"unsupported backend type %q", backendType)
+		return nil, status.Errorf(codes.InvalidArgument, "unsupported backend type %q", backendType)
 	}
 
 	// If we have a workspace specified, get the value for that. Use the default otherwise
 	workspaceName := be.DefaultStateName
-	if workspacePB, hasWorkspaceTypePB := req.GetProperties().GetFields()["workspace"]; hasWorkspaceTypePB {
-		if _, isString := workspacePB.Kind.(*structpb.Value_StringValue); !isString {
-			return nil, status.Errorf(codes.InvalidArgument,
-				"expected a string value for property %q", "workspace")
-		}
-
-		workspaceName = workspacePB.GetStringValue()
-	}
-
-	// Convert our backendConfig struct to something usable with the backend configuration schema
-	terraformNamedNews := structpbNamesPulumiToTerraform(req.GetProperties())
-
-	// Delete fields which are known not to be part of the Terraform remote state config
-	// which will cause unknown field failures if passed to a go-cty CoerceValue function
-	delete(terraformNamedNews.GetFields(), "backend_type")
-	delete(terraformNamedNews.GetFields(), "workspace")
-	delete(terraformNamedNews.GetFields(), "outputs")
-
-	backendConfigCty, err := structpbToCtyObject(terraformNamedNews)
-	if err != nil {
-		return nil, status.Errorf(codes.Internal, "error mapping config from Pulumi format to cty: %s", err)
+	if args.Workspaces.Name != "" {
+		workspaceName = args.Workspaces.Name
 	}
 
 	// Get the configuration schema from the backend
 	backend := backendInitFn()
 
 	// Attempt to coerce our config object into the config schema types - note errors
-	backendConfigCoerced, err := backend.ConfigSchema().CoerceValue(backendConfigCty)
+	backendConfigCoerced, err := backend.ConfigSchema().CoerceValue(cty.ObjectVal(map[string]cty.Value{
+		"token":         cty.StringVal(args.BackendConfig.Token),
+		"organization":  cty.StringVal(args.BackendConfig.Organization),
+		"hostname":      cty.StringVal(args.BackendConfig.Hostname),
+		"resource_name": cty.StringVal(args.BackendConfig.ResourceName),
+	}))
 	if err != nil {
 		return nil, status.Errorf(codes.Internal, "error coercing config from Pulumi format to cty: %s", err)
 	}
@@ -105,24 +111,17 @@ func RemoteStateReferenceRead(ctx context.Context, req *pulumirpc.ReadRequest) (
 		return nil, status.Error(codes.NotFound, "remote state not found")
 	}
 
-	// Get the root module outputs and process them from a map of string to cty.Value into a structpb
-	outputsCty := state.RootModule().OutputValues
-	outputsStructpb, err := outputsToStructpb(outputsCty)
-	if err != nil {
-		return nil, status.Errorf(codes.Internal, "error converting Terraform outputs: %s", err)
+	var outputs map[string]any
+	for k, v := range state.RootModule().OutputValues {
+		jsonBytes, err := ctyjson.Marshal(v.Value, v.Value.Type())
+		if err != nil {
+			return nil, fmt.Errorf("error marshaling cty to JSON: %w", err)
+		}
+		var goV any
+		if err := json.Unmarshal(jsonBytes, &goV); err != nil {
+			return nil, fmt.Errorf("error unmarshaling JSON: %w", err)
+		}
+		outputs[k] = goV
 	}
-
-	// Construct our properties based on the outputs
-	req.GetProperties().Fields["outputs"] = &structpb.Value{
-		Kind: &structpb.Value_StructValue{
-			StructValue: outputsStructpb,
-		},
-	}
-
-	// Return a successful response to the engine
-	return &pulumirpc.ReadResponse{
-		Id:         req.Id,
-		Inputs:     req.Inputs,
-		Properties: req.GetProperties(),
-	}, nil
+	return outputs, nil
 }
